TASK: TASKS.md:2110-2132 - libc/libsys 64-bit syscall wrappers
- Added ftruncate/truncate wrappers with 64-bit length split
- Verified lseek already correctly splits 64-bit offset
- Fixed sys_time declaration to int64_t in perso_native.c
- Added SYS_truncate/SYS_ftruncate to userspace syscall.h

Build verified: kernel and libc compile without errors.

+++ sys/exec/perso/perso_native.c
@@ -34,7 +34,7 @@ extern int sys_getdents(unsigned int, void*, unsigned int);
 struct thr_param;
 extern int sys_thr_new(struct thr_param*, int);
 extern int sys_acct(const char*);
-extern int sys_time(uint32_t*);
+extern int64_t sys_time(int64_t*);
 extern int sys_getpid(void);
 extern int sys_getcwd(char*, size_t);
 extern int sys_execve(const char*, char**, char**);
